Upgrade to latest versions of Spring and Spring Security.

+++ src/main/java/edu/hawaii/its/casdemo/service/ActionService.java
@@ -1,13 +1,14 @@
 package edu.hawaii.its.casdemo.service;
 
-import edu.hawaii.its.casdemo.type.Action;
-import edu.hawaii.its.casdemo.type.ActionLog;
 import java.util.List;
 
+import edu.hawaii.its.casdemo.type.Action;
+import edu.hawaii.its.casdemo.type.ActionLog;
+
 public interface ActionService {
-    public List<Action> findActions();
+	public List<Action> findActions();
 
-    public void record(ActionLog actionLog);
+	public void record(ActionLog actionLog);
 
-    public int logCount();
+	public long logCount();
 }
